Add load progress

+++ Mount3DX/Form1.cs
@@ -18,8 +18,8 @@ namespace Mount3DX
             InitializeComponent();
         }
 
-        public static string PROGRAM_NAME = "Mount 3DX";
-        public static string PROGRAM_VERSION = "0.1";
+        public static readonly string PROGRAM_NAME = "Mount 3DX";
+        public static readonly string PROGRAM_VERSION = "0.1";
 
         private void Form1_Load(object sender, EventArgs e)
         {
@@ -27,88 +27,18 @@ namespace Mount3DX
 
             LoadSettings();
 
-            Scratch();
+            //Scratch();
 
-            //todo: The Windows 10 WebDAV client sometimes restricts access to files larger than 50MB.
-            //This can be changed in: HKEY_LOCAL_MACHINE\SYSTEM\CurrentControlSet\Services\WebClient\Parameters\FileSizeLimitInBytes = 4294967295
-            //After which, the 'WebClient' service needs to be restarted.
+            //todo: The Windows 10 WebDAV client needs several settings tweaked:
+            //HKEY_LOCAL_MACHINE\SYSTEM\CurrentControlSet\Services\WebClient\Parameters
+            //      FileAttributesLimitInBytes = 4294967295     //Allows explorer to open folders which contain many items
+            //      FileSizeLimitInBytes = 4294967295           //Allows large files to be opened
+            //      InternetServerTimeoutInSec = 180            //Give the server a longer time to respond
+            //      LocalServerTimeoutInSec = 180               //Give the server a longer time to respond
+            //      SendReceiveTimeoutInSec = 180               //Give the server a longer time to respond
+            //After changing the settings, restart the 'WebClient' service
 
             lblRunningStatus.Text = string.Empty;
-        }
-
-        private void Scratch()
-        {
-            var _3dxServer = new _3dxServer(settings._3dx.ServerUrl, settings._3dx.CookiesString, false, 5);
-
-            var root = new _3dxFolder()
-            {
-                Name = ""
-            };
-            var roots = _3dxServer.GetRootFolders();
-            root.Subfolders.AddRange(roots);
-
-            var securityContext = _3dxServer.GetSecurityContext();
-
-            /*
-            roots
-                .Recurse(folder =>
-                {
-                    var itemsInFolder = _3dxServer.GetItemsInFolder(folder, securityContext);
-
-                    var documents = itemsInFolder
-                                        .OfType<_3dxDocument>()
-                                        .ToList();
-
-                    folder.Subfolders = itemsInFolder
-                                        .Except(documents)
-                                        .OfType<_3dxFolder>()
-                                        .ToList();
-
-                    Debug.WriteLine($"{folder.FullPath}\tDocuments: {documents.Count:N0}\tSubfolders: {folder.Subfolders.Count:N0}");
-
-                    return folder.Subfolders;
-                })
-                .ToList();
-            */
-
-
-            var folderQueue = new ConcurrentQueue<_3dxFolder>();
-            roots.ForEach(rt => folderQueue.Enqueue(rt));
-
-            int totalFolders = 0;
-            int totalDocs = 0;
-            int totalFiles = 0;
-
-            var recurseTask = QueueUtility
-                    .Process(folderQueue, folder =>
-                    {
-                        var itemsInFolder = _3dxServer.GetItemsInFolder(folder, securityContext);
-
-                        var documents = itemsInFolder
-                                            .OfType<_3dxDocument>()
-                                            .ToList();
-
-                        var files = documents
-                                        .Sum(doc => doc.Files.Count);
-
-                        folder.Subfolders = itemsInFolder
-                                            .Except(documents)
-                                            .OfType<_3dxFolder>()
-                                            .ToList();
-
-
-
-                        Interlocked.Add(ref totalFolders, folder.Subfolders.Count);
-                        Interlocked.Add(ref totalDocs, documents.Count);
-                        Interlocked.Add(ref totalFiles, files);
-
-                        //Debug.WriteLine($"{folder.FullPath}\tSubfolders: {folder.Subfolders.Count:N0}\tDocs: {documents.Count:N0}");
-                        Debug.WriteLine($"Total folders: {totalFolders:N0}\tTotal docs: {totalDocs:N0}\tTotal files: {totalFiles:N0}");
-
-                        return folder.Subfolders;
-
-                    }, settings._3dx.QueryThreads, new CancellationToken());
-            recurseTask.Wait();
         }
 
         readonly string settingsFilename = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "settings.json");
@@ -143,7 +73,9 @@ namespace Mount3DX
                 settings.Vfs.MapToDriveLetter = txtMapToDriveLetter.Text;
 
                 var settingsJson = settings.SerializeToJson();
-                if (!string.IsNullOrEmpty(settingsJson))
+
+                var existingSettingsFileContent = File.ReadAllText(settingsJson);
+                if (!string.IsNullOrEmpty(settingsJson) && !settingsJson.Equals(existingSettingsFileContent))
                 {
                     File.WriteAllText(settingsFilename, settingsJson);
                 }
@@ -158,9 +90,9 @@ namespace Mount3DX
 
         Session? session = null;
 
-        private void btnStart_Click(object sender, EventArgs e)
+        private void BtnStart_Click(object sender, EventArgs e)
         {
-            lblRunningStatus.BackColor = Control.DefaultBackColor;
+            lblRunningStatus.BackColor = DefaultBackColor;
             lblRunningStatus.ForeColor = Color.Black;
             lblRunningStatus.Text = "";
 
@@ -186,7 +118,7 @@ namespace Mount3DX
                             break;
 
                         case ProgressEventArgs.EnumNature.Neutral:
-                            lblRunningStatus.BackColor = Control.DefaultBackColor;
+                            lblRunningStatus.BackColor = DefaultBackColor;
                             lblRunningStatus.ForeColor = Color.Black;
                             break;
 
@@ -228,6 +160,74 @@ namespace Mount3DX
                 grp3dx.Enabled = true;
                 grpVfs.Enabled = true;
             }
+        }
+
+        private void Scratch()
+        {
+            var _3dxServer = new _3dxServer(settings._3dx.ServerUrl, settings._3dx.CookiesString, false, 5);
+
+            var root = new _3dxFolder(
+                                "root",
+                                "",
+                                null,
+                                DateTime.Now,
+                                DateTime.Now,
+                                DateTime.Now)
+            {
+                Subfolders = _3dxServer.GetRootFolders()
+            };
+
+            var securityContext = _3dxServer.GetSecurityContext();
+
+            var folderQueue = new ConcurrentQueue<_3dxFolder>();
+            root.Subfolders.ForEach(rt => folderQueue.Enqueue(rt));
+
+            int totalFolders = 0;
+            int totalDocs = 0;
+            int totalFiles = 0;
+
+            var recurseTask = QueueUtility
+                    .Process(folderQueue, folder =>
+                    {
+                        var itemsInFolder = _3dxServer.GetItemsInFolder(folder, securityContext);
+
+                        var documents = itemsInFolder
+                                            .OfType<_3dxDocument>()
+                                            .ToList();
+
+                        var files = documents
+                                        .Sum(doc => doc.Files.Count);
+
+                        folder.Subfolders = itemsInFolder
+                                            .Except(documents)
+                                            .OfType<_3dxFolder>()
+                                            .ToList();
+
+                        Interlocked.Add(ref totalFolders, folder.Subfolders.Count);
+                        Interlocked.Add(ref totalDocs, documents.Count);
+                        Interlocked.Add(ref totalFiles, files);
+
+                        //Debug.WriteLine($"{folder.FullPath}\tSubfolders: {folder.Subfolders.Count:N0}\tDocs: {documents.Count:N0}");
+                        Debug.WriteLine($"Total folders: {totalFolders:N0}\tTotal docs: {totalDocs:N0}\tTotal files: {totalFiles:N0}");
+
+                        return folder.Subfolders;
+
+                    }, settings._3dx.QueryThreads, new CancellationToken());
+            recurseTask.Wait();
+
+            var itemTypes = _3dxServer
+                                .itemTypes
+                                .GroupBy(
+                                    itemType => itemType,
+                                    itemType => itemType,
+                                    (key, grp) => new
+                                    {
+                                        ItemType = key,
+                                        Count = grp.Count()
+                                    })
+                                .OrderByDescending(grp => grp.Count)
+                                .Select(grp => $"{grp.ItemType},{grp.Count}")
+                                .ToString(Environment.NewLine);
         }
     }
 }
